Adjust import/export logic on business-hours block
